// highlight-next/src/graph/generated/operations.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  Any: any;
  Int64: any;
  Timestamp: any;
};

export type BackendErrorObjectInput = {
  event: Scalars['String'];
  payload?: InputMaybe<Scalars['String']>;
  request_id: Scalars['String'];
  session_secure_id: Scalars['String'];
  source: Scalars['String'];
  stackTrace: Scalars['String'];
  timestamp: Scalars['Timestamp'];
  type: Scalars['String'];
  url: Scalars['String'];
};
// ...
export type PushBackendPayloadMutationVariables = Types.Exact<{
  errors: Array<Types.InputMaybe<Types.BackendErrorObjectInput>> | Types.InputMaybe<Types.BackendErrorObjectInput>;
}>;
// ...
export type Sdk = ReturnType<typeof getSdk>;